FEATURE ICON AND HEADER
Created IconSvg component this component serves to render the icons in the format SVG, which icon that must be shown is passed via prop on it which queries a svg of svg.
The following tests were created for him:
 - Should that the component return svg
 - Should that the component return svg the value of name passed the prop

Created the Header component of the application, this component only receives within a logo of the project.
To ensure the quality of the project, the following tests were created for this component:
 - Should component Header contain div header__logo
 - Should return the SVG with name soon inside the logo div

We added the lib node-sass to be able to use SASS to write the CSS and the enzyme-adapter-react-15 enzyme-adapter-react-16 lib to probe enzyme