SHOWCASE
Added the react router call in the project.
A home view was created and within that view the search and showcase component was placed.

In the search component the onSubmit function was done, which when pisquisa hits the URL of the imdb api and returns the result that is placed in the state showcase.

In the showcase component it was passed via the props state showcase which contains the api return, made some CSS adjustments that was breaking into the responsive and changed the position of the dark mask into the tag figure to correct an error of getting a space without mascara.

We created tests for the showcase component:

  - Should wrapper contain image equal to the past in the date Poster field
  - Should the wrapper component contain a title equal to the last one by date
  - Should the wrapper component contain a year equal to the last in the data
  - Should the wrapper component contain the Link with ID equal to the last in the imdbID field in the data
  - Should wrapper return two li
  - Should wrapper2 contain two image equal to the past in the data2 Poster field
  - Should the wrapper2 component contain two Title equal to the last one by data2
  - Should the wrapper2 component contain Two Year equal to the last in the data2
  - Should the wrapper component contain two Link with ID equal to the last in the imdbID field in the data2