FEATURE SEARCH
Placed a validation for when trying to do a search without putting a movie is put a red embroider around the input and does not get in the API.

Set the type parameter in the GET url to compute only movies.

Removed a file / which I had created unintentionally and was disrupting I download the project in another machine.