FEATURE SHOWCASE
Created a loader for while doing this get in the api of a feedback to the user that something is happening.

Made the validation for when you do not find any video it shows the message "Do not know what to search?" with an image.

The following tests were created:

- Should return class hide in loader if passed to prop with loaderShowcase as hide
- Should return class hide in empty if passed to prop with movieEmpty as hide