FEATURE SHOWCASE / MOVIE TITLE
200~Placed a click function in the window, and when a movie is
clicked this function gets a get into the imDB api and returns the movie
information and puts that information into a state.

Changed the movie title component to receive the dynamic value that was
passed via the movie state.

The following tests were created for the movie-title component:

- Should MovieTitle return one h1
- Should MovieTitle return the component with text Movie passed via prop
- Should MovieTitle return the component with default text that is title
when not passing the prop