FEATURE BACK
Changed the Babk component, now it has a link that redirects the page to home.
When you click this link it calls a function and this function clears the state of the showcase and movie component.
In app.js and I removed the part that cleaned the state of the movie.