FEATURE MOVIE / SHOWCASE
Placing the componentDidMount in the app for when it comes directly to the movie page it takes the movie id in the url and gets it in the api and fills in the state to appear the information on the page.

I created a component for when the get result in the api is empty it shows this component in the window and the movie page.

For this component the following tests were created:
- Should Empty return one empty
- Should return the component with class hide passad via prop