FEATURE FAVOURITES
Created the function that adds the movie to favorites.
This function takes the id of the movie and saves it in the localstorage, every time it clicks on a movie or reloads the movie page it goes in the localstorage and checks if it has the ID.
If it has the ID it is called the one state of the application that makes change the visual of the button to the format added.