Included the lib dotenv-webpack in the project.
With lib we can create an .env file which contains all the information about the api keys and keys.

I got one to fail in the persistence of the favorite button, when it did not think it was breaking the code, a simple validation was placed to check if it is not null.